feat: Allow tag style "original" for additional tags
Some edge cases require using the tag style from avro as-is
for other formats.
The tag style `original` explicitly does not do any processing
on the tag before code generation.

This is mostly already supported in the code generation,
but the CLI options would reject
any configuration value which would trigger this behaviour.