feat: allow passing in custom options for OCF decoder.
Currently, the OCF Decoder will fail when parsing OCF files that contain
a field that exceeds the default limit of 1MiB. There is currently no
way to configure this limit for the OCF Decoder.

This change provides a way to construct an OCF Decoder using a custom
avro.API, which enables configuration of the decoder.