feat: add SOE codecs
Avro describes a very simple Single Object Encoding in
https://avro.apache.org/docs/1.12.0/specification/#single-object-encoding-specification

Encoders put a schema fingerprint in the header, which decoders can use to look
up a schema definition for deserialization.

There are two codecs:

* soe.Codec: follows avro.Marshal/Unmarshal but adds/parses SOE header
* soe.TypedCodec: a type-safe generic version for avrogen-generated types

Both of them can encode/decode SOE, using a single schema.

Also add soe.DynamicDecoder, which has a more liberal view of types and schemas:
for each record, try to look up the associated schema from a configured
resolver, and decode anything for which a schema is found. This can be used for
full dynamic encoding of streams of records from different encoders, as long as
their schemas are available for decoding. This is similar to how the Avro Java
implementation does it.

Include a minimal in-memory schema store.